Update copyright, switch resume view to TLS
Updates the copyright to the new year and fixes resume
display. Previously the Google PDF viewer was using an URL that was not
TLS and was referencing Google Docs. At some point Google stopped
serving requests for non-TLS ans changed the URL to reference Google
Drive. The embedding URL had to be changed as a result.